requirements.txt
pyyaml>=6.0
streamlit>=1.30.0
flask>=2.0.0


